feat(oversight): show that iOS CWVs are fake
As of 2024, CWV readings with iOS user agents are coming from non-iOS browsers
that falsify their user agent string. This can include Chrome when using devtools
and simulating a mobile viewport. Apple does not allow custom rendering engines
in most of the world, and in the EU, where it does allow custom rendering engines,
no browser vendor ships a browser with a non-WebKit engine, so that all actual
iOS browsers are just WebKit in disguise. Webkit does not implement key APIs required
for measurement of Core Web Vitals, effectively rendering all CWVs readings that
claim to apply to iOS suspect.

This PR addresses this by reducing the visual contrast of the CWV pills for the
user agent iOS only.